requirements.txt
dj-database-url==0.5.0
Django==2.2.4
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.10.2
gunicorn==19.9.0
psycopg2==2.8.3
pytz==2019.2
sqlparse==0.3.0
whitenoise==4.1.3
